backend/requirements.txt
flask
flask-cors
numpy
scipy
gunicorn
gunicorn
gunicorn
gunicorn
google-genai
